Moved df_to_* functions to their own file, updated pytests and cli.py
